move websocket types to types.ts
