+++ frontend/src/app/pages/devices-components/picow-power-button.ts
@@ -2,8 +2,8 @@ import { power as svgPower } from "ui/svg/smoothie-line-icons";
 
 import { html, UIIconButton } from "ui";
 import * as api from "../../../lib/api";
+import type { WSEvents_Device } from "../../../lib/websocket";
 import ws from "../../../lib/websocket";
-import type { WSEvents_Device } from "../../../lib/websocket/ws-events";
 import type { PicowStore } from "../../../types";
 export type PicowPowerButton_States = "active" | "pending" | null;
