+++ frontend/src/lib/websocket/index.ts
@@ -1,3 +1,4 @@
+export * from "./types";
 import { WSEvents } from "./ws-events";
 
 export default new WSEvents();
